+++ src/index.js
@@ -22,18 +22,29 @@ export class ErrorBoundary extends Component {
   }
 }
 
-export const createFetcher = (fetchFunc) => {
-  const refetch = (input) => {
+const isPromise = x => typeof x.then === 'function';
+
+const createRunFetch = (fetchFunc) => {
+  const runFetch = (input) => {
     const state = { pending: true };
-    state.promise = (async () => {
+    const start = async () => {
       try {
         state.data = await fetchFunc(input);
       } catch (e) {
-        state.error = e;
+        if (isPromise(e)) {
+          try {
+            await e;
+          } finally {
+            await start();
+          }
+        } else {
+          state.error = e;
+        }
       } finally {
         state.pending = false;
       }
-    })();
+    };
+    state.promise = start();
     return {
       get data() {
         if (state.pending) throw state.promise;
@@ -41,17 +52,22 @@ export const createFetcher = (fetchFunc) => {
         return state.data;
       },
       get refetch() {
-        return refetch;
+        return runFetch;
       },
     };
   };
+  return runFetch;
+};
+
+export const createFetcher = (fetchFunc) => {
+  const runFetch = createRunFetch(fetchFunc);
   return {
-    prefetch: input => refetch(input),
+    prefetch: input => runFetch(input),
     lazyFetch: (fallbackData) => {
       let suspendable = null;
       const fetchOnce = (input) => {
         if (!suspendable) {
-          suspendable = refetch(input);
+          suspendable = runFetch(input);
         }
         return suspendable;
       };
@@ -72,15 +88,15 @@ export const createFetcher = (fetchFunc) => {
 
 export const useSuspendable = (suspendable) => {
   const [result, setResult] = useState(suspendable);
-  const origFetch = suspendable.refetch;
+  const origRefetch = suspendable.refetch;
   return {
     get data() {
       return result.data;
     },
     refetch: useCallback((nextInput) => {
-      const nextResult = origFetch(nextInput);
+      const nextResult = origRefetch(nextInput);
       setResult(nextResult);
-    }, [origFetch]),
+    }, [origRefetch]),
     lazy: result.lazy,
   };
 };